# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: P2P Cloud Gaming signaling, session, and wallet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hosts/register:
    post:
      operationId: registerHost
      tags: [hosts]
      summary: Register a new host
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterHostBody"
      responses:
        "201":
          description: Host registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
  /hosts/{hostToken}:
    get:
      operationId: getHost
      tags: [hosts]
      summary: Look up host by token
      parameters:
        - in: path
          name: hostToken
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hosts/{hostToken}/config:
    patch:
      operationId: updateHostConfig
      tags: [hosts]
      summary: Update the host's offer config (binding, pricing, schedule, restream)
      parameters:
        - in: path
          name: hostToken
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHostConfigBody"
      responses:
        "200":
          description: Updated host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Host not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hosts/{hostToken}/sessions:
    get:
      operationId: listHostSessions
      tags: [hosts]
      summary: List sessions for a host
      parameters:
        - in: path
          name: hostToken
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "404":
          description: Host not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hosts/{hostToken}/stats:
    get:
      operationId: getHostStats
      tags: [hosts]
      summary: Aggregate stats for a host (sessions, minutes streamed, earnings)
      parameters:
        - in: path
          name: hostToken
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Host stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostStats"
        "404":
          description: Host not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hosts/{hostToken}/activity:
    get:
      operationId: getHostActivity
      tags: [hosts]
      summary: Recent activity feed for a host
      parameters:
        - in: path
          name: hostToken
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activity items, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
        "404":
          description: Host not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /players/register:
    post:
      operationId: registerPlayer
      tags: [players]
      summary: Register a new player wallet identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPlayerBody"
      responses:
        "201":
          description: Player registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{playerToken}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Look up player by token
      parameters:
        - in: path
          name: playerToken
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ScheduleSlot:
      type: object
      description: A single weekly availability window (UTC).
      properties:
        day:
          type: integer
          minimum: 0
          maximum: 6
          description: 0 = Sunday … 6 = Saturday
        startMin:
          type: integer
          minimum: 0
          maximum: 1440
        endMin:
          type: integer
          minimum: 0
          maximum: 1440
      required:
        - day
        - startMin
        - endMin
    Host:
      type: object
      properties:
        id:
          type: string
        hostToken:
          type: string
        displayName:
          type: string
        internalBalanceLzt:
          type: integer
          description: Синий — internal LZT, cannot be withdrawn
        withdrawableBalanceLzt:
          type: integer
          description: Зелёный — LZT convertible back to crypto at 200:1
        gameId:
          type: ["string", "null"]
          description: Catalog game this host is bound to
        boundAppPath:
          type: string
          description: Absolute Windows path to the .exe the agent will launch
        boundUrl:
          type: string
          description: URL of a browser game the agent will open. When set, takes precedence over boundAppPath.
        boundAppLabel:
          type: string
          description: Friendly label shown in the games library
        description:
          type: string
        tags:
          type: array
          items:
            type: string
          description: Capability tags shown as badges and used as library filters
        launchPriceUsd:
          type: number
          description: Charged once when a player joins (may be negative)
        minutePriceUsd:
          type: number
          description: Charged per minute while streaming (may be negative)
        scheduleMode:
          type: string
          enum: [always, scheduled]
        scheduleJson:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleSlot"
        streamPlatform:
          type: string
          description: e.g. "twitch", "youtube", "rtmp"
        streamUrl:
          type: string
        streamKeySet:
          type: boolean
          description: True if a stream key is stored. The key itself is never returned.
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
      required:
        - id
        - hostToken
        - displayName
        - internalBalanceLzt
        - withdrawableBalanceLzt
        - gameId
        - boundAppPath
        - boundUrl
        - boundAppLabel
        - description
        - tags
        - launchPriceUsd
        - minutePriceUsd
        - scheduleMode
        - scheduleJson
        - streamPlatform
        - streamUrl
        - streamKeySet
        - createdAt
        - lastSeenAt
    UpdateHostConfigBody:
      type: object
      description: Partial update — omit a field to leave it unchanged.
      properties:
        gameId:
          type: ["string", "null"]
        boundAppPath:
          type: string
        boundUrl:
          type: string
          description: URL of a browser game; mutually exclusive with boundAppPath at runtime.
        boundAppLabel:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        launchPriceUsd:
          type: number
        minutePriceUsd:
          type: number
        scheduleMode:
          type: string
          enum: [always, scheduled]
        scheduleJson:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleSlot"
        streamPlatform:
          type: string
        streamUrl:
          type: string
        streamKey:
          type: string
          description: Pass empty string to clear the stored key.
    Player:
      type: object
      properties:
        id:
          type: string
        playerToken:
          type: string
        displayName:
          type: string
        internalBalanceLzt:
          type: integer
          description: Синий — internal LZT, cannot be withdrawn
        withdrawableBalanceLzt:
          type: integer
          description: Зелёный — LZT convertible back to crypto at 200:1
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
      required:
        - id
        - playerToken
        - displayName
        - internalBalanceLzt
        - withdrawableBalanceLzt
        - createdAt
        - lastSeenAt
    RegisterHostBody:
      type: object
      properties:
        displayName:
          type: string
      required:
        - displayName
    RegisterPlayerBody:
      type: object
      properties:
        displayName:
          type: string
      required:
        - displayName
...
    Session:
      type: object
      properties:
        id:
          type: string
        hostId:
          type: string
        playerToken:
          type: string
        claimedByPlayerId:
          type: ["string", "null"]
        appName:
          type: string
        status:
          type: string
          description: One of pending, active, ended
        resolution:
          type: string
        bitrateKbps:
          type: integer
        ratePerMinute:
          type: number
          description: Player credits charged per minute (host receives net of commission)
        paymentSource:
          type: string
          enum: [blue, green, auto]
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - hostId
        - playerToken
        - appName
        - status
        - resolution
        - bitrateKbps
        - ratePerMinute
        - createdAt
...
    HostStats:
      type: object
      properties:
        totalSessions:
          type: integer
        activeSessions:
          type: integer
        totalMinutesStreamed:
          type: integer
        lifetimeEarnings:
          type: number
        earnings7d:
          type: number
        internalBalanceLzt:
          type: integer
        withdrawableBalanceLzt:
          type: integer
      required:
        - totalSessions
        - activeSessions
        - totalMinutesStreamed
        - lifetimeEarnings
        - earnings7d
        - internalBalanceLzt
        - withdrawableBalanceLzt
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - kind
        - title
        - timestamp
...