# lib/api-zod/src/generated/types/host.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * P2P Cloud Gaming signaling, session, and wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { HostScheduleMode } from "./hostScheduleMode";
import type { ScheduleSlot } from "./scheduleSlot";

export interface Host {
  id: string;
  hostToken: string;
  displayName: string;
  /** Синий — internal LZT, cannot be withdrawn */
  internalBalanceLzt: number;
  /** Зелёный — LZT convertible back to crypto at 200:1 */
  withdrawableBalanceLzt: number;
  /**
   * Catalog game this host is bound to
   * @nullable
   */
  gameId: string | null;
  /** Absolute Windows path to the .exe the agent will launch */
  boundAppPath: string;
  /** URL of a browser game the agent will open. When set, takes precedence over boundAppPath. */
  boundUrl: string;
  /** Friendly label shown in the games library */
  boundAppLabel: string;
  description: string;
  /** Capability tags shown as badges and used as library filters */
  tags: string[];
  /** Charged once when a player joins (may be negative) */
  launchPriceUsd: number;
  /** Charged per minute while streaming (may be negative) */
  minutePriceUsd: number;
  scheduleMode: HostScheduleMode;
  scheduleJson: ScheduleSlot[];
  /** e.g. "twitch", "youtube", "rtmp" */
  streamPlatform: string;
  streamUrl: string;
  /** True if a stream key is stored. The key itself is never returned. */
  streamKeySet: boolean;
  createdAt: Date;
  lastSeenAt: Date;
}