# lib/api-zod/src/generated/types/hostStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * P2P Cloud Gaming signaling, session, and wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface HostStats {
  totalSessions: number;
  activeSessions: number;
  totalMinutesStreamed: number;
  lifetimeEarnings: number;
  earnings7d: number;
  internalBalanceLzt: number;
  withdrawableBalanceLzt: number;
}