# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * P2P Cloud Gaming signaling, session, and wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: string;
  playerToken: string;
  displayName: string;
  /** Синий — internal LZT, cannot be withdrawn */
  internalBalanceLzt: number;
  /** Зелёный — LZT convertible back to crypto at 200:1 */
  withdrawableBalanceLzt: number;
  createdAt: Date;
  lastSeenAt: Date;
}